+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-alpha79] - unreleased
+- handling element max height in markdown
+- refactoring item preparation, making sure datatables works in Quarto in the new setup
+
 ## [2-alpha78] - 2024-02-04
 - using Quarto raw-html and data-qmd support (thus supporting formulae in tables) - fixing #54
 - showing keywords correctly in column names - fixing #52